pyproject.toml
[project]
name = "mineria-de-datos-proyecto-examen"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "xlrd (>=2.0.1,<3.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "selenium (>=4.30.0,<5.0.0)",
    "pdfplumber (>=0.11.6,<0.12.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "flask (>=3.1.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
